seperate log.py file
