+++ kowalski/ingesters/ingest_ptf_matchfiles.py
@@ -18,10 +18,10 @@ from kowalski.utils import (
     deg2dms,
     deg2hms,
     init_db_sync,
-    log,
     Mongo,
 )
 from kowalski.config import load_config
+from kowalski.log import log
 
 
 """ load config and secrets """
