+++ kowalski/ingesters/ingest_turbo.py
@@ -16,10 +16,10 @@ from kowalski.utils import (
     deg2dms,
     deg2hms,
     init_db_sync,
-    log,
     Mongo,
 )
 from kowalski.config import load_config
+from kowalski.log import log
 
 
 """ load config and secrets """
