+++ kowalski/tests/test_ingester_pgir.py
@@ -7,14 +7,14 @@ import requests
 from kowalski.alert_brokers.alert_broker_pgir import watchdog
 from kowalski.ingesters.ingester import KafkaStream
 from test_ingester_ztf import Program, Filter
-from kowalski.utils import Mongo, init_db_sync, log
+from kowalski.utils import Mongo, init_db_sync
 from kowalski.config import load_config
+from kowalski.log import log
 
 """ load config and secrets """
-
-USING_DOCKER = os.environ.get("USING_DOCKER", False)
 config = load_config(config_files=["config.yaml"])["kowalski"]
 
+USING_DOCKER = os.environ.get("USING_DOCKER", False)
 if USING_DOCKER:
     config["server"]["host"] = "kowalski_api_1"
