+++ kowalski/tests/test_ingester_wntr.py
@@ -6,8 +6,9 @@ import time
 from kowalski.alert_brokers.alert_broker_winter import watchdog
 from kowalski.ingesters.ingester import KafkaStream
 from test_ingester_ztf import Program, Filter
-from kowalski.utils import Mongo, init_db_sync, log
+from kowalski.utils import Mongo, init_db_sync
 from kowalski.config import load_config
+from kowalski.log import log
 
 """
 Essentially the same as the ZTF ingester tests (test_ingester_ztf.py), except -
