+++ kowalski/utils.py
@@ -19,7 +19,6 @@ __all__ = [
     "init_db_sync",
     "jd_to_date",
     "jd_to_datetime",
-    "log",
     "memoize",
     "mjd_to_datetime",
     "Mongo",
@@ -41,7 +40,6 @@ import hashlib
 import inspect
 import io
 import math
-import os
 import secrets
 import string
 import sys
@@ -61,6 +59,8 @@ from motor.motor_asyncio import AsyncIOMotorClient
 from numba import jit
 from pymongo.errors import BulkWriteError
 from requests.adapters import HTTPAdapter
+
+from kowalski.log import time_stamp, log
 
 LOG_DIR = "./logs"
 
@@ -105,27 +105,6 @@ class TimeoutHTTPAdapter(HTTPAdapter):
             return super().send(request, **kwargs)
         except AttributeError:
             kwargs["timeout"] = DEFAULT_TIMEOUT
-
-
-def time_stamp():
-    """
-
-    :return: UTC time -> string
-    """
-    return datetime.datetime.utcnow().strftime("%Y%m%d_%H:%M:%S")
-
-
-def log(message):
-    timestamp = time_stamp()
-    print(f"{timestamp}: {message}")
-
-    if not os.path.isdir(LOG_DIR):
-        os.mkdir(LOG_DIR)
-
-    date = timestamp.split("_")[0]
-    with open(os.path.join(LOG_DIR, f"kowalski_{date}.log"), "a") as logfile:
-        logfile.write(f"{timestamp}: {message}\n")
-        logfile.flush()
 
 
 def forgiving_true(expression):
